refactor: split ssg html rendering
