+++ crates/ox_content_ssg/src/html.rs
@@ -1,17 +1,22 @@
 //! HTML page generation for SSG.
 
 use askama::Template;
-use memchr::memmem;
 
+mod entry;
+mod footer;
 mod nav;
 mod page;
+mod render;
+mod social;
 mod theme;
+mod theme_css;
+mod utils;
 
-use nav::generate_nav_html;
 pub use page::{
     EntryPageConfig, FeatureConfig, HeroAction, HeroConfig, HeroImage, HeroNoticeConfig,
     LocaleInfo, NavGroup, NavItem, PageData, SsgConfig, TocEntry,
 };
+pub use render::{generate_bare_html, generate_html};
 pub use theme::{
     SocialLink, SocialLinks, ThemeColors, ThemeConfig, ThemeEmbed, ThemeEntryPage, ThemeFonts,
     ThemeFooter, ThemeHeader, ThemeLayout,
@@ -166,1013 +171,5 @@ const SSG_JS: &str = include_str!("ssg.js");
 /// it is inert for the default no-JavaScript tab widget.
 const TABS_JS: &str = include_str!("plugins/tabs.js");
 
-/// Generates CSS variable overrides for theme colors.
-fn generate_theme_css(theme: &ThemeConfig) -> String {
-    let mut css = String::new();
-
-    // Light mode colors
-    if let Some(ref colors) = theme.colors {
-        let mut vars = Vec::new();
-        if let Some(ref v) = colors.primary {
-            vars.push(format!("--octc-color-primary: {v};"));
-        }
-        if let Some(ref v) = colors.primary_hover {
-            vars.push(format!("--octc-color-primary-hover: {v};"));
-        }
-        if let Some(ref v) = colors.background {
-            vars.push(format!("--octc-color-bg: {v};"));
-        }
-        if let Some(ref v) = colors.background_alt {
-            vars.push(format!("--octc-color-bg-alt: {v};"));
-        }
-        if let Some(ref v) = colors.text {
-            vars.push(format!("--octc-color-text: {v};"));
-        }
-        if let Some(ref v) = colors.text_muted {
-            vars.push(format!("--octc-color-text-muted: {v};"));
-        }
-        if let Some(ref v) = colors.border {
-            vars.push(format!("--octc-color-border: {v};"));
-        }
-        if let Some(ref v) = colors.code_background {
-            vars.push(format!("--octc-color-code-bg: {v};"));
-        }
-        if let Some(ref v) = colors.code_text {
-            vars.push(format!("--octc-color-code-text: {v};"));
-        }
-        if !vars.is_empty() {
-            css.push_str(":root {\n  ");
-            css.push_str(&vars.join("\n  "));
-            css.push_str("\n}\n");
-        }
-    }
-
-    // Dark mode colors
-    if let Some(ref colors) = theme.dark_colors {
-        let mut vars = Vec::new();
-        if let Some(ref v) = colors.primary {
-            vars.push(format!("--octc-color-primary: {v};"));
-        }
-        if let Some(ref v) = colors.primary_hover {
-            vars.push(format!("--octc-color-primary-hover: {v};"));
-        }
-        if let Some(ref v) = colors.background {
-            vars.push(format!("--octc-color-bg: {v};"));
-        }
-        if let Some(ref v) = colors.background_alt {
-            vars.push(format!("--octc-color-bg-alt: {v};"));
-        }
-        if let Some(ref v) = colors.text {
-            vars.push(format!("--octc-color-text: {v};"));
-        }
-        if let Some(ref v) = colors.text_muted {
-            vars.push(format!("--octc-color-text-muted: {v};"));
-        }
-        if let Some(ref v) = colors.border {
-            vars.push(format!("--octc-color-border: {v};"));
-        }
-        if let Some(ref v) = colors.code_background {
-            vars.push(format!("--octc-color-code-bg: {v};"));
-        }
-        if let Some(ref v) = colors.code_text {
-            vars.push(format!("--octc-color-code-text: {v};"));
-        }
-        if !vars.is_empty() {
-            css.push_str("[data-theme=\"dark\"] {\n  ");
-            css.push_str(&vars.join("\n  "));
-            css.push_str("\n}\n");
-            css.push_str("@media (prefers-color-scheme: dark) {\n  :root:not([data-theme=\"light\"]) {\n    ");
-            css.push_str(&vars.join("\n    "));
-            css.push_str("\n  }\n}\n");
-        }
-    }
-
-    // Layout overrides
-    if let Some(ref layout) = theme.layout {
-        let mut vars = Vec::new();
-        if let Some(ref v) = layout.sidebar_width {
-            vars.push(format!("--octc-sidebar-width: {v};"));
-        }
-        if let Some(ref v) = layout.header_height {
-            vars.push(format!("--octc-header-height: {v};"));
-        }
-        if let Some(ref v) = layout.max_content_width {
-            vars.push(format!("--octc-max-content-width: {v};"));
-        }
-        if !vars.is_empty() {
-            css.push_str(":root {\n  ");
-            css.push_str(&vars.join("\n  "));
-            css.push_str("\n}\n");
-        }
-    }
-
-    // Font overrides
-    if let Some(ref fonts) = theme.fonts {
-        let mut vars = Vec::new();
-        if let Some(ref v) = fonts.sans {
-            vars.push(format!("--octc-font-sans: {v};"));
-        }
-        if let Some(ref v) = fonts.mono {
-            vars.push(format!("--octc-font-mono: {v};"));
-        }
-        if !vars.is_empty() {
-            css.push_str(":root {\n  ");
-            css.push_str(&vars.join("\n  "));
-            css.push_str("\n}\n");
-        }
-    }
-
-    // Custom CSS
-    if let Some(ref custom_css) = theme.css {
-        css.push_str(custom_css);
-    }
-
-    css
-}
-
-/// Generates footer HTML from theme configuration.
-fn generate_footer_html(theme: &ThemeConfig) -> String {
-    let footer = match &theme.footer {
-        Some(f) if f.message.is_some() || f.copyright.is_some() => f,
-        _ => return String::new(),
-    };
-
-    let template = FooterTemplate {
-        message: footer.message.as_deref(),
-        copyright: footer.copyright.as_deref(),
-    };
-    template.render().unwrap_or_default()
-}
-
-/// Converts a `.md` link to an HTML path for entry page frontmatter links.
-/// Entry pages are always `index.md`, so relative links like `getting-started.md`
-/// become `{base}getting-started/index.html`.
-fn convert_entry_link(link: &str, base: &str) -> String {
-    // Don't touch external URLs or anchor-only links
-    if link.starts_with("http://") || link.starts_with("https://") || link.starts_with('#') {
-        return link.to_string();
-    }
-
-    // Split into path and fragment
-    let (path, fragment) = match link.split_once('#') {
-        Some((p, f)) => (p, Some(f)),
-        None => (link, None),
-    };
-
-    let is_md =
-        std::path::Path::new(path).extension().is_some_and(|ext| ext.eq_ignore_ascii_case("md"));
-
-    if !is_md {
-        return link.to_string();
-    }
-
-    // Remove .md extension
-    let stem = &path[..path.len() - 3];
-
-    // Entry page is always index.md, so plain relative: getting-started.md -> {base}getting-started/index.html
-    let converted = if stem == "index" || stem.ends_with("/index") {
-        let dir = stem.trim_end_matches("/index").trim_end_matches("index");
-        if dir.is_empty() {
-            format!("{base}index.html")
-        } else {
-            format!("{base}{dir}/index.html")
-        }
-    } else {
-        format!("{base}{stem}/index.html")
-    };
-
-    match fragment {
-        Some(f) => format!("{converted}#{f}"),
-        None => converted,
-    }
-}
-
-/// Generates the Entry page HTML (hero section and features).
-fn generate_entry_html(entry: &EntryPageConfig, base: &str) -> String {
-    // Convert hero config to view
-    let hero_view = entry.hero.as_ref().map(|hero| {
-        let actions = hero.actions.as_ref().map(|actions| {
-            actions
-                .iter()
-                .map(|action| {
-                    let theme_class = match action.theme.as_deref() {
-                        Some("brand") | None => "hero-action-brand",
-                        Some("alt") => "hero-action-alt",
-                        _ => "hero-action-brand",
-                    };
-                    let href = convert_entry_link(&action.link, base);
-                    HeroActionView {
-                        href,
-                        theme_class: theme_class.to_string(),
-                        text: action.text.clone(),
-                    }
-                })
-                .collect()
-        });
-
-        // Process hero image src
-        let image = hero.image.as_ref().map(|img| {
-            let src = convert_entry_link(&img.src, base);
-            let light_src = img.light_src.as_ref().map(|src| convert_entry_link(src, base));
-            let dark_src = img.dark_src.as_ref().map(|src| convert_entry_link(src, base));
-            HeroImage {
-                src,
-                light_src,
-                dark_src,
-                alt: img.alt.clone(),
-                width: img.width,
-                height: img.height,
-            }
-        });
-
-        HeroView {
-            name: hero.name.clone(),
-            text: hero.text.clone(),
-            tagline: hero.tagline.clone(),
-            notice: hero.notice.clone(),
-            image,
-            actions,
-        }
-    });
-
-    // Convert features config to view
-    let features_view: Option<Vec<FeatureView>> = entry.features.as_ref().map(|features| {
-        features
-            .iter()
-            .map(|feature| {
-                let has_link = feature.link.is_some();
-                let tag = if has_link { "a" } else { "div" };
-                let href_attr = feature
-                    .link
-                    .as_ref()
-                    .map(|link| {
-                        let href = convert_entry_link(link, base);
-                        format!(" href=\"{href}\"")
-                    })
-                    .unwrap_or_default();
-
-                let icon_html = feature.icon.as_ref().map(|icon| render_icon(icon, base));
-
-                FeatureView {
-                    tag,
-                    href_attr,
-                    icon_html,
-                    title: feature.title.clone(),
-                    details: feature.details.clone(),
-                    has_link,
-                }
-            })
-            .collect()
-    });
-
-    let template = EntryTemplate { hero: hero_view.as_ref(), features: features_view.as_deref() };
-    template.render().unwrap_or_default()
-}
-
-/// Footer CSS styles (added when footer is used).
-const FOOTER_CSS: &str = r"
-.site-footer {
-  margin-top: 3rem;
-  padding: 2rem 1.5rem;
-  border-top: 1px solid var(--octc-color-border);
-  text-align: center;
-  color: var(--octc-color-text-muted);
-  font-size: 0.875rem;
-}
-.site-footer p {
-  margin: 0.25rem 0;
-}
-.site-footer a {
-  color: var(--octc-color-primary);
-}
-.site-footer a:hover {
-  color: var(--octc-color-primary-hover);
-}
-";
-
-fn wrap_css_section(name: &str, css: &str) -> String {
-    if css.trim().is_empty() {
-        return String::new();
-    }
-
-    // These markers let `externalize_shared_page_assets` split one generated
-    // `<style>` block into stable shared chunks. They are CSS comments, so the
-    // page remains valid even before the extraction pass runs.
-    format!("/* ox-content:css:{name}:start */\n{css}\n/* ox-content:css:{name}:end */\n")
-}
-
-fn page_content_contains_any(content: &str, needles: &[&str]) -> bool {
-    // Page asset detection checks for a small set of HTML markers in the final
-    // content. `memmem` keeps each probe on optimized byte search instead of
-    // Rust's substring machinery, and the caller short-circuits as soon as one
-    // marker is present.
-    let haystack = content.as_bytes();
-    needles.iter().any(|needle| memmem::find(haystack, needle.as_bytes()).is_some())
-}
-
-fn escape_html(value: &str) -> String {
-    // SSG escaping favors predictable allocation over chained `replace()`:
-    // allocate once at the input length, stream characters into the output,
-    // and let `String` grow only if replacements make the result longer.
-    let mut output = String::with_capacity(value.len());
-    for ch in value.chars() {
-        match ch {
-            '&' => output.push_str("&amp;"),
-            '<' => output.push_str("&lt;"),
-            '>' => output.push_str("&gt;"),
-            '"' => output.push_str("&quot;"),
-            '\'' => output.push_str("&#39;"),
-            _ => output.push(ch),
-        }
-    }
-    output
-}
-
-fn generate_toc_html(toc: &[TocEntry]) -> String {
-    let mut html = String::new();
-
-    for entry in toc {
-        let depth = entry.depth.clamp(1, 6);
-        html.push_str("        <li class=\"toc-item\"><a href=\"#");
-        html.push_str(&escape_html(&entry.slug));
-        html.push_str("\" class=\"toc-link toc-link--depth-");
-        html.push_str(&depth.to_string());
-        html.push_str("\">");
-        html.push_str(&escape_html(&entry.text));
-        html.push_str("</a></li>\n");
-    }
-
-    html
-}
-
-fn civil_from_days(days: i64) -> (i64, i64, i64) {
-    let z = days + 719_468;
-    let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = z - era * 146_097;
-    let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096) / 365;
-    let year = yoe + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let day = doy - (153 * mp + 2) / 5 + 1;
-    let month = mp + if mp < 10 { 3 } else { -9 };
-    let year = year + i64::from(month <= 2);
-    (year, month, day)
-}
-
-fn format_last_updated(timestamp_ms: i64) -> Option<LastUpdatedView> {
-    if timestamp_ms < 0 {
-        return None;
-    }
-
-    let days = (timestamp_ms / 1_000).div_euclid(86_400);
-    let (year, month, day) = civil_from_days(days);
-    let date = format!("{year:04}-{month:02}-{day:02}");
-    Some(LastUpdatedView { text: date.clone(), datetime: date })
-}
-
-fn infer_locale_dir(locale: &str) -> &'static str {
-    match locale.split('-').next().unwrap_or_default().to_ascii_lowercase().as_str() {
-        "ar" | "fa" | "he" | "iw" | "ps" | "sd" | "ug" | "ur" | "yi" => "rtl",
-        _ => "ltr",
-    }
-}
-
-fn html_locale_attrs(config: &SsgConfig) -> (&str, &str) {
-    let lang = config
-        .locale
-        .as_deref()
-        .filter(|locale| !locale.trim().is_empty())
-        .or_else(|| config.available_locales.as_ref()?.first().map(|locale| locale.code.as_str()))
-        .unwrap_or("en");
-    let configured_dir = config.available_locales.as_ref().and_then(|locales| {
-        locales.iter().find(|locale| locale.code == lang).and_then(|locale| {
-            match locale.dir.as_str() {
-                "ltr" => Some("ltr"),
-                "rtl" => Some("rtl"),
-                _ => None,
-            }
-        })
-    });
-
-    (lang, configured_dir.unwrap_or_else(|| infer_locale_dir(lang)))
-}
-
-/// Generates a complete HTML page for SSG.
-///
-/// This function creates a full HTML document with navigation sidebar,
-/// content area, table of contents, search functionality, and theme toggle.
-pub fn generate_html(page_data: &PageData, nav_groups: &[NavGroup], config: &SsgConfig) -> String {
-    let nav_html = generate_nav_html(nav_groups, &page_data.path);
-
-    // Theme configuration
-    let theme = config.theme.as_ref();
-    let embed = theme.and_then(|t| t.embed.as_ref());
-
-    // Generate theme CSS overrides
-    let theme_css = theme.map_or(String::new(), generate_theme_css);
-
-    // Check if we have a footer
-    let has_footer = theme.is_some_and(|t| {
-        t.footer.as_ref().is_some_and(|f| f.message.is_some() || f.copyright.is_some())
-    });
-    let footer_css = if has_footer { FOOTER_CSS } else { "" };
-
-    // Check if this is an entry page
-    let is_entry_page = page_data.entry_page.is_some();
-    // Build CSS as named sections instead of one anonymous blob. Shared,
-    // content-addressed extraction can then pull out only the sections that are
-    // globally cacheable and keep page-specific or relative-url CSS inline.
-    let mut css_sections = vec![wrap_css_section("base", SSG_CSS)];
-
-    if is_entry_page {
-        css_sections.push(wrap_css_section("entry", ENTRY_CSS));
-    }
-    if page_content_contains_any(&page_data.content, &["ox-tabs", "ox-tab-panel"]) {
-        css_sections.push(wrap_css_section("plugin-tabs", TABS_CSS));
-    }
-    if page_content_contains_any(&page_data.content, &["ox-youtube"]) {
-        css_sections.push(wrap_css_section("plugin-youtube", YOUTUBE_CSS));
-    }
-    if page_content_contains_any(
-        &page_data.content,
-        &["ox-github-card", "ox-github-code", "ox-github-error"],
-    ) {
-        css_sections.push(wrap_css_section("plugin-github", GITHUB_CSS));
-    }
-    if page_content_contains_any(&page_data.content, &["ox-ogp-card", "ox-ogp-simple"]) {
-        css_sections.push(wrap_css_section("plugin-ogp", OGP_CSS));
-    }
-    if page_content_contains_any(&page_data.content, &["ox-mermaid"]) {
-        css_sections.push(wrap_css_section("plugin-mermaid", MERMAID_CSS));
-    }
-    if page_content_contains_any(&page_data.content, &["data-ox-island", "ox-island"]) {
-        css_sections.push(wrap_css_section("plugin-island", ISLAND_CSS));
-    }
-    if has_footer {
-        css_sections.push(wrap_css_section("footer", footer_css));
-    }
-    if !theme_css.is_empty() {
-        css_sections.push(wrap_css_section("theme", &theme_css));
-    }
-
-    let all_css = css_sections.join("");
-    let toc_html = generate_toc_html(&page_data.toc);
-    let has_toc = !toc_html.is_empty();
-    let last_updated = page_data.last_updated.and_then(format_last_updated);
-
-    // Embedded HTML for specific positions
-    let embed_head = embed.and_then(|e| e.head.as_deref()).unwrap_or("");
-    let embed_header_before = embed.and_then(|e| e.header_before.as_deref()).unwrap_or("");
-    let embed_header_after = embed.and_then(|e| e.header_after.as_deref()).unwrap_or("");
-    let embed_sidebar_before = embed.and_then(|e| e.sidebar_before.as_deref()).unwrap_or("");
-    let embed_sidebar_after = embed.and_then(|e| e.sidebar_after.as_deref()).unwrap_or("");
-    let embed_content_before = embed.and_then(|e| e.content_before.as_deref()).unwrap_or("");
-    let embed_content_after = embed.and_then(|e| e.content_after.as_deref()).unwrap_or("");
-    let embed_footer_before = embed.and_then(|e| e.footer_before.as_deref()).unwrap_or("");
-
-    // Footer HTML
-    let footer_html = if let Some(embed_footer) = embed.and_then(|e| e.footer.clone()) {
-        embed_footer
-    } else if let Some(t) = theme {
-        generate_footer_html(t)
-    } else {
-        String::new()
-    };
-
-    // Header logo customization
-    let header_config = theme.and_then(|t| t.header.as_ref());
-    let logo_url = header_config
-        .and_then(|h| h.logo.as_ref())
-        .map_or_else(|| "logo.svg", std::string::String::as_str);
-    let logo_width = header_config.and_then(|h| h.logo_width).unwrap_or(28);
-    let logo_height = header_config.and_then(|h| h.logo_height).unwrap_or(28);
-    let show_site_name_text = header_config.and_then(|h| h.show_site_name_text).unwrap_or(true);
-
-    let resolve_theme_asset = |url: &str| {
-        if url.starts_with("http://") || url.starts_with("https://") || url.starts_with('/') {
-            url.to_string()
-        } else {
-            format!("{}{}", config.base, url)
-        }
-    };
-
-    // Build logo src (prepend base if not absolute URL)
-    let logo_src = resolve_theme_asset(logo_url);
-    let logo_light_src =
-        header_config.and_then(|h| h.logo_light.as_deref()).map(resolve_theme_asset);
-    let logo_dark_src = header_config.and_then(|h| h.logo_dark.as_deref()).map(resolve_theme_asset);
-
-    // Custom JS
-    let custom_js = theme.and_then(|t| t.js.as_deref()).unwrap_or("");
-    let all_js =
-        format!("{}\n{}\n{}", SSG_JS.replace("{{base}}", &config.base), TABS_JS, custom_js);
-
-    // Social links
-    let social_links_html = theme
-        .and_then(|t| t.social_links.as_ref())
-        .map_or(String::new(), generate_social_links_html);
-
-    // Mobile footer social links
-    let mobile_social_links_html = theme
-        .and_then(|t| t.social_links.as_ref())
-        .map_or(String::new(), generate_mobile_social_links_html);
-
-    // Generate entry page content if applicable
-    let (page_class, main_content) = if let Some(ref entry) = page_data.entry_page {
-        let entry_html = generate_entry_html(entry, &config.base);
-        // Entry page: hero/features + optional markdown content
-        let combined = if page_data.content.trim().is_empty() {
-            entry_html
-        } else {
-            format!(
-                "{}\n<div class=\"entry-content\">\n  <div class=\"content\">\n{}\n  </div>\n</div>",
-                entry_html, page_data.content
-            )
-        };
-        ("entry-page", combined)
-    } else {
-        ("", format!("<article class=\"content\">\n{}\n      </article>", page_data.content))
-    };
-
-    let mut body_classes = Vec::new();
-    if !page_class.is_empty() {
-        body_classes.push(page_class.to_string());
-    }
-    if is_entry_page
-        && theme.and_then(|t| t.entry_page.as_ref()).and_then(|entry| entry.mode.as_deref())
-            == Some("subtle")
-    {
-        body_classes.push("entry-page--subtle".to_string());
-    }
-    let body_class = body_classes.join(" ");
-
-    let document_title = if page_data.title.trim() == config.site_name.trim() {
-        config.site_name.clone()
-    } else {
-        format!("{} - {}", page_data.title, config.site_name)
-    };
-    let (html_lang, html_dir) = html_locale_attrs(config);
-
-    let template = PageTemplate {
-        html_lang,
-        html_dir,
-        site_name: &config.site_name,
-        document_title: &document_title,
-        description: page_data.description.as_deref(),
-        og_image: config.og_image.as_deref(),
-        css: &all_css,
-        embed_head,
-        body_class: &body_class,
-        embed_header_before,
-        embed_header_after,
-        base: &config.base,
-        logo_src: &logo_src,
-        logo_light_src: logo_light_src.as_deref(),
-        logo_dark_src: logo_dark_src.as_deref(),
-        show_site_name_text,
-        logo_width,
-        logo_height,
-        social_links: &social_links_html,
-        is_entry_page,
-        embed_sidebar_before,
-        navigation: &nav_html,
-        embed_sidebar_after,
-        embed_content_before,
-        main_content: &main_content,
-        has_toc,
-        toc_html: &toc_html,
-        last_updated: last_updated.as_ref(),
-        embed_content_after,
-        embed_footer_before,
-        footer_html: &footer_html,
-        mobile_social_links: &mobile_social_links_html,
-        js: &all_js,
-    };
-
-    template.render().unwrap_or_default()
-}
-
-/// Generates a bare HTML page for SSG.
-///
-/// This page intentionally omits navigation, styles, and scripts.
-pub fn generate_bare_html(content: &str, title: &str) -> String {
-    BarePageTemplate { title, content }.render().unwrap_or_default()
-}
-
-/// Renders an icon based on its format.
-///
-/// Supported formats:
-/// - `mdi:icon-name` - Material Design Icons via Iconify CDN
-/// - `lucide:icon-name` - Lucide icons via Iconify CDN
-/// - `{prefix}:{name}` - Any Iconify icon set
-/// - URL (http://, https://) - Direct image URL
-/// - Path ending with .svg, .png - Local image path
-/// - Other - Treated as emoji/text
-fn render_icon(icon: &str, base: &str) -> String {
-    // Check for Iconify format (prefix:name)
-    if let Some((prefix, name)) = icon.split_once(':') {
-        // Validate it looks like an icon reference (not a URL scheme)
-        if !prefix.contains('/') && !name.starts_with("//") {
-            // Convert to Iconify CDN URL
-            let iconify_url = format!("https://api.iconify.design/{prefix}/{name}.svg");
-            // Use span with mask-image for color control
-            return format!(
-                "<span class=\"iconify-icon\" style=\"-webkit-mask-image: url('{iconify_url}'); mask-image: url('{iconify_url}')\"></span>"
-            );
-        }
-    }
-
-    // Check if it's an image URL
-    if icon.starts_with("http://") || icon.starts_with("https://") {
-        return format!("<img src=\"{icon}\" alt=\"\" />");
-    }
-
-    // Check if it's a local image path
-    if icon.ends_with(".svg") || icon.ends_with(".png") {
-        let icon_src =
-            if icon.starts_with('/') { icon.to_string() } else { format!("{base}{icon}") };
-        return format!("<img src=\"{icon_src}\" alt=\"\" />");
-    }
-
-    // Treat as emoji/text
-    icon.to_string()
-}
-
-fn generate_social_links_html(links: &SocialLinks) -> String {
-    let template = SocialLinksTemplate {
-        github: links.github.as_deref(),
-        twitter: links.twitter.as_deref(),
-        discord: links.discord.as_deref(),
-    };
-    let mut html = template.render().unwrap_or_default();
-    if let Some(custom_links) = &links.links {
-        for link in custom_links {
-            html.push_str(&render_custom_social_link(link, false));
-        }
-    }
-    html
-}
-
-fn generate_mobile_social_links_html(links: &SocialLinks) -> String {
-    let template = MobileSocialLinksTemplate {
-        github: links.github.as_deref(),
-        twitter: links.twitter.as_deref(),
-        discord: links.discord.as_deref(),
-    };
-    let mut html = template.render().unwrap_or_default();
-    if let Some(custom_links) = &links.links {
-        for link in custom_links {
-            html.push_str(&render_custom_social_link(link, true));
-        }
-    }
-    html
-}
-
-fn render_custom_social_link(link: &SocialLink, mobile: bool) -> String {
-    let label = link.aria_label.as_deref().or(link.icon.as_deref()).unwrap_or("Social link");
-    let href = escape_html(&link.link);
-    let label = escape_html(label);
-    let icon_html = render_social_icon(link);
-
-    if mobile {
-        format!(
-            "<a href=\"{href}\" class=\"mobile-footer-btn\" aria-label=\"{label}\" target=\"_blank\" rel=\"noopener\">{icon_html}<span class=\"mobile-footer-label\">{label}</span></a>\n"
-        )
-    } else {
-        format!(
-            "<a href=\"{href}\" class=\"social-link\" aria-label=\"{label}\" target=\"_blank\" rel=\"noopener\">{icon_html}</a>\n"
-        )
-    }
-}
-
-fn render_social_icon(link: &SocialLink) -> String {
-    if let Some(svg) = link.icon_svg.as_deref().and_then(validate_social_svg) {
-        return svg.to_string();
-    }
-
-    link.icon
-        .as_deref()
-        .map(|icon| format!("<span class=\"social-link-icon\">{}</span>", escape_html(icon)))
-        .unwrap_or_default()
-}
-
-fn validate_social_svg(svg: &str) -> Option<&str> {
-    let trimmed = svg.trim();
-    let lower = trimmed.to_ascii_lowercase();
-    if trimmed.starts_with("<svg") && !lower.contains("<script") {
-        Some(trimmed)
-    } else {
-        None
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_generate_html() {
-        let page_data = PageData {
-            title: "Test Page".to_string(),
-            description: Some("Test description".to_string()),
-            content: "<h1>Hello</h1>".to_string(),
-            toc: vec![TocEntry { depth: 1, text: "Hello".to_string(), slug: "hello".to_string() }],
-            last_updated: Some(0),
-            path: "test".to_string(),
-            entry_page: None,
-        };
-
-        let nav_groups = vec![NavGroup {
-            title: "Guide".to_string(),
-            items: vec![NavItem {
-                title: "Test Page".to_string(),
-                path: "test".to_string(),
-                href: "/docs/test/index.html".to_string(),
-                children: vec![],
-                collapsed: None,
-                sticky_collapsed: None,
-            }],
-            collapsed: None,
-            sticky_collapsed: None,
-        }];
-
-        let config = SsgConfig {
-            site_name: "Test Site".to_string(),
-            base: "/docs/".to_string(),
-            og_image: None,
-            theme: None,
-            locale: None,
-            available_locales: None,
-        };
-
-        let html = generate_html(&page_data, &nav_groups, &config);
-
-        assert!(html.contains("Test Page - Test Site"));
-        assert!(html.contains("<h1>Hello</h1>"));
-        assert!(html.contains("Guide"));
-        assert!(html.contains("class=\"toc\""));
-        assert!(html.contains("href=\"#hello\""));
-        assert!(html.contains("Last updated:"));
-        assert!(html.contains("<time datetime=\"1970-01-01\">1970-01-01</time>"));
-    }
-
-    #[test]
-    fn test_generate_bare_html() {
-        let html = generate_bare_html("<h1>Hello</h1>", "Test Page");
-
-        assert_eq!(
-            html,
-            "<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n  <meta charset=\"UTF-8\">\n  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n  <title>Test Page</title>\n</head>\n<body>\n<h1>Hello</h1>\n</body>\n</html>"
-        );
-    }
-
-    #[test]
-    fn test_generate_bare_html_escapes_title_but_keeps_content_raw() {
-        let html = generate_bare_html("<h1>Raw & ready</h1>", "<script>alert(1)</script>");
-        let title = html
-            .split("<title>")
-            .nth(1)
-            .and_then(|value| value.split("</title>").next())
-            .expect("bare HTML should include a title element");
-
-        assert!(title.contains("alert(1)"));
-        assert!(!title.contains('<'));
-        assert!(!title.contains('>'));
-        assert!(html.contains("<h1>Raw & ready</h1>"));
-        assert!(!html.contains("<title><script>"));
-    }
-
-    #[test]
-    fn test_generate_nav_html_with_nested_collapsed_items() {
-        let nav_groups = vec![NavGroup {
-            title: "Guide & API".to_string(),
-            collapsed: Some(true),
-            sticky_collapsed: Some(true),
-            items: vec![NavItem {
-                title: "Runtime <Core>".to_string(),
-                path: "runtime".to_string(),
-                href: "javascript:alert(1)".to_string(),
-                collapsed: Some(false),
-                sticky_collapsed: Some(true),
-                children: vec![NavItem {
-                    title: "Setup".to_string(),
-                    path: "runtime/setup".to_string(),
-                    href: "/docs/runtime/setup/index.html".to_string(),
-                    children: vec![],
-                    collapsed: None,
-                    sticky_collapsed: None,
-                }],
-            }],
-        }];
-
-        let html = generate_nav_html(&nav_groups, "runtime/setup");
-        assert!(html.contains(
-            "<details class=\"nav-section nav-section--collapsible\" data-ox-nav-state-key=\"group:0:Guide &amp; API\">"
-        ));
-        assert!(html.contains(
-            "<details class=\"nav-details\" open data-ox-nav-state-key=\"item:group:0.0:runtime:Runtime &lt;Core&gt;\">"
-        ));
-        assert!(html.contains("Guide &amp; API"));
-        assert!(html.contains("Runtime &lt;Core&gt;"));
-        assert!(html.contains("href=\"#\""));
-        assert!(html.contains("nav-list nav-list--nested"));
-        assert!(html.contains("class=\"nav-link active\""));
-        assert!(!html.contains("javascript:"));
-    }
-
-    #[test]
-    fn test_generate_html_with_theme() {
-        let page_data = PageData {
-            title: "Themed Page".to_string(),
-            description: None,
-            content: "<p>Content</p>".to_string(),
-            toc: vec![],
-            last_updated: None,
-            path: "themed".to_string(),
-            entry_page: None,
-        };
-
-        let nav_groups = vec![];
-
-        let config = SsgConfig {
-            site_name: "Themed Site".to_string(),
-            base: "/".to_string(),
-            og_image: None,
-            locale: None,
-            available_locales: None,
-            theme: Some(ThemeConfig {
-                colors: Some(ThemeColors {
-                    primary: Some("#3498db".to_string()),
-                    ..Default::default()
-                }),
-                footer: Some(ThemeFooter {
-                    message: Some("Built with ox-content".to_string()),
-                    copyright: Some("2025 Test".to_string()),
-                }),
-                ..Default::default()
-            }),
-        };
-
-        let html = generate_html(&page_data, &nav_groups, &config);
-
-        // Check theme CSS is applied
-        assert!(html.contains("--octc-color-primary: #3498db;"));
-        // Check footer is present
-        assert!(html.contains("Built with ox-content"));
-        assert!(html.contains("2025 Test"));
-    }
-
-    #[test]
-    fn test_generate_html_with_custom_social_link() {
-        let page_data = PageData {
-            title: "Social Page".to_string(),
-            description: None,
-            content: "<p>Content</p>".to_string(),
-            toc: vec![],
-            last_updated: None,
-            path: "social".to_string(),
-            entry_page: None,
-        };
-        let config = SsgConfig {
-            site_name: "Social Site".to_string(),
-            base: "/".to_string(),
-            og_image: None,
-            locale: None,
-            available_locales: None,
-            theme: Some(ThemeConfig {
-                social_links: Some(SocialLinks {
-                    links: Some(vec![SocialLink {
-                        icon: None,
-                        icon_svg: Some("<svg viewBox=\"0 0 24 24\"></svg>".to_string()),
-                        link: "https://example.com".to_string(),
-                        aria_label: Some("Example".to_string()),
-                    }]),
-                    ..Default::default()
-                }),
-                ..Default::default()
-            }),
-        };
-
-        let html = generate_html(&page_data, &[], &config);
-
-        assert!(html.contains("aria-label=\"Example\""));
-        assert!(html.contains("href=\"https://example.com\""));
-        assert!(html.contains("<svg viewBox=\"0 0 24 24\"></svg>"));
-        assert!(html.contains("<span class=\"mobile-footer-label\">Example</span>"));
-    }
-
-    #[test]
-    fn test_generate_html_without_toc_omits_outline() {
-        let page_data = PageData {
-            title: "No TOC".to_string(),
-            description: None,
-            content: "<p>Content</p>".to_string(),
-            toc: vec![],
-            last_updated: None,
-            path: "no-toc".to_string(),
-            entry_page: None,
-        };
-        let config = SsgConfig {
-            site_name: "Test Site".to_string(),
-            base: "/".to_string(),
-            og_image: None,
-            theme: None,
-            locale: None,
-            available_locales: None,
-        };
-
-        let html = generate_html(&page_data, &[], &config);
-
-        assert!(!html.contains("class=\"toc\""));
-        assert!(!html.contains("<main class=\"main main--with-toc\">"));
-        assert!(!html.contains("Last updated:"));
-    }
-
-    #[test]
-    fn test_format_last_updated_rejects_invalid_timestamps() {
-        assert!(format_last_updated(-1).is_none());
-    }
-
-    #[test]
-    fn test_html_locale_attrs_use_current_locale_and_direction() {
-        let config = SsgConfig {
-            site_name: "Localized".to_string(),
-            base: "/".to_string(),
-            og_image: None,
-            theme: None,
-            locale: Some("ar".to_string()),
-            available_locales: None,
-        };
-
-        assert_eq!(html_locale_attrs(&config), ("ar", "rtl"));
-
-        let page_data = PageData {
-            title: "مرحبا".to_string(),
-            description: None,
-            content: "<p>Content</p>".to_string(),
-            toc: vec![],
-            last_updated: None,
-            path: "ar".to_string(),
-            entry_page: None,
-        };
-        let html = generate_html(&page_data, &[], &config);
-        assert!(html.contains("<html lang=\"ar\" dir=\"rtl\">"));
-    }
-
-    #[test]
-    fn test_generate_toc_html_escapes_entries() {
-        let html = generate_toc_html(&[TocEntry {
-            depth: 2,
-            text: "A <script>".to_string(),
-            slug: "a\" onclick=\"alert(1)".to_string(),
-        }]);
-
-        assert!(html.contains("A &lt;script&gt;"));
-        assert!(html.contains("href=\"#a&quot; onclick=&quot;alert(1)\""));
-        assert!(!html.contains("A <script>"));
-    }
-
-    #[test]
-    fn test_generate_theme_css() {
-        let theme = ThemeConfig {
-            colors: Some(ThemeColors {
-                primary: Some("#ff0000".to_string()),
-                background: Some("#ffffff".to_string()),
-                ..Default::default()
-            }),
-            dark_colors: Some(ThemeColors {
-                primary: Some("#ff6666".to_string()),
-                ..Default::default()
-            }),
-            layout: Some(ThemeLayout {
-                sidebar_width: Some("300px".to_string()),
-                ..Default::default()
-            }),
-            ..Default::default()
-        };
-
-        let css = generate_theme_css(&theme);
-
-        assert!(css.contains("--octc-color-primary: #ff0000;"));
-        assert!(css.contains("--octc-color-bg: #ffffff;"));
-        assert!(css.contains("[data-theme=\"dark\"]"));
-        assert!(css.contains("--octc-sidebar-width: 300px;"));
-    }
-
-    #[test]
-    fn test_generate_footer_html() {
-        let theme = ThemeConfig {
-            footer: Some(ThemeFooter {
-                message: Some("Footer message".to_string()),
-                copyright: Some("Copyright info".to_string()),
-            }),
-            ..Default::default()
-        };
-
-        let html = generate_footer_html(&theme);
-
-        assert!(html.contains("site-footer"));
-        assert!(html.contains("Footer message"));
-        assert!(html.contains("Copyright info"));
-    }
-}
+mod tests;
