+++ crates/ox_content_ssg/src/html/tests.rs
@@ -1,0 +1,2 @@
+mod rendering;
+mod theme;
